+++ metadata.php
@@ -32,7 +32,7 @@ $aModule = [
     'extend' => [
         \OxidEsales\Eshop\Application\Model\OrderArticle::class => \FreshAdvance\Invoice\Document\Model\OrderArticleExtension::class,
         \OxidEsales\Eshop\Core\Language::class => \FreshAdvance\Invoice\Language\Extension\Language::class,
-        \OxidEsales\Eshop\Core\Email::class => \FreshAdvance\Invoice\Transition\Core\Email::class,
+        \OxidEsales\Eshop\Core\Email::class => \FreshAdvance\Invoice\Email\Core\EmailExtension::class,
     ],
     'settings' => [
         /** Main */