+++ tests/Integration/Document/Model/OrderArticleExtensionTest.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace FreshAdvance\Invoice\Tests\Unit\Transition\Model;
+namespace FreshAdvance\Invoice\Tests\Integration\Document\Model;
 
 use FreshAdvance\Invoice\Document\Model\OrderArticleExtension;
 use OxidEsales\Eshop\Core\Model\BaseModel;
@@ -32,7 +32,10 @@ class OrderArticleExtensionTest extends IntegrationTestCase
         ]);
         $testProduct->save();
 
-        $orderArticle = $this->createPartialMock(OrderArticleExtension::class, ['getParentId', 'getProductId']);
+        $orderArticle = $this->createPartialMock(
+            OrderArticleExtension::class,
+            ['getParentId', 'getProductId']
+        );
         $orderArticle->method('getParentId')->willReturn('');
         $orderArticle->method('getProductId')->willReturn($testProductId);
 
@@ -65,7 +68,10 @@ class OrderArticleExtensionTest extends IntegrationTestCase
         ]);
         $testProduct->save();
 
-        $orderArticle = $this->createPartialMock(\FreshAdvance\Invoice\Document\Model\OrderArticleExtension::class, ['getParentId', 'getProductId']);
+        $orderArticle = $this->createPartialMock(
+            \FreshAdvance\Invoice\Document\Model\OrderArticleExtension::class,
+            ['getParentId', 'getProductId']
+        );
         $orderArticle->method('getParentId')->willReturn($testParentProductId);
         $orderArticle->method('getProductId')->willReturn($testProductId);
 
@@ -98,7 +104,10 @@ class OrderArticleExtensionTest extends IntegrationTestCase
         ]);
         $testProduct->save();
 
-        $orderArticle = $this->createPartialMock(\FreshAdvance\Invoice\Document\Model\OrderArticleExtension::class, ['getParentId', 'getProductId']);
+        $orderArticle = $this->createPartialMock(
+            \FreshAdvance\Invoice\Document\Model\OrderArticleExtension::class,
+            ['getParentId', 'getProductId']
+        );
         $orderArticle->method('getParentId')->willReturn($testParentProductId);
         $orderArticle->method('getProductId')->willReturn($testProductId);